[10.18] add epvm
[10.18] add epvm

+++ node/node.go
@@ -5,10 +5,10 @@ type Node struct {
 	Ip              string
 	Api             string
 	Cores           int
-	Memory          int
-	MemoryAllocated int
-	Disk            int
-	DiskAllocated   int
+	Memory          int64
+	MemoryAllocated int64
+	Disk            int64
+	DiskAllocated   int64
 	Role            string
 	TaskCount       int
 }
